requirements.txt
django==4.2.1
petl==1.7.12
requests==2.31.0
django-environ==0.10.0
psycopg2==2.9.6
pytest==7.3.1
httpx==0.24.1
uvicorn==0.22.0
pytest-asyncio==0.21.0
factory-boy==3.2.1
pytest-django==4.5.2
